EV-221: Att Crypto Coin List and create Chart Model
