+++ lib/Widgets/model_charts.dart
@@ -1,0 +1,5 @@
+class CapitalDate {
+  CapitalDate(this.days, this.marketCapital);
+  final DateTime days;
+  final int marketCapital;
+}
